Removed client storage of cookies
Client no longer needs to save the username of themselves. Only the UUID given by the server. This is easier on the server, no descrepencies from client to server, as well as allowing for more control over the sessions.

Also added a "User has left chat" message on a disconnection from chat.